Added --config=crafty
